Alter cosign module to allow local user creation.
The cosign module disables the functionality to create users locally.
But we need to add a user to the Drupal user table to satisfy the
requirements of the roles module so we can assign varying
access privileges to authenticated users.

The changes are taken from Bob Riddle's working notes. The current
changes were taken from
http://www-personal.umich.edu/~bdr/et/swback/umms-cosign.html#alter_cosign.module_to_permit_add_user