Second part of commit for OERDRUPAL-63
Logged in admin users can now see unpublished courses, resources and
subunits in grey. This was done as a set of very ugly tempate file
hacks that inspect the "node_status" value and apply the
".unpublished" css class to all items that have status 0 which
has them rendered in grey according to the defined css rules.