OERDRUPAL-70 http/https inconsistencies
A base_url setting of https://[...] was added to fix an issue with autocomplete.
This causes normal unauthenticated users to get redirected into https:// when
they needn't be.  Completely removing the base_url setting fixes the issue
with autocomplete.

Completely removing the base_url setting brought up a couple of other
issues with images not being found.  Fix the references to these images
to match the others (with leading slashes).