Delete settings.php symlink in default site.
sites/default/settings.php is a symlink to
sites/oer.switchbackcms.com/settings.php which doesn't exist. Having
this symlink in the repository was preventing the .gitignore directive to
ignore the sites/default/settings.php file from working. So we removed
the sites/default/settings.php symlink from the respository.