Ignore .htaccess and rename to .htaccess-template.
The .htaccess file is environment specific so we shouldn't blanket
install a .htaccess file. Now, we ignore the .htaccess file and have
provided a .htaccess-template file that can be used to set up a working
.htaccess file.

The rewrite directives for cosign and other applications on the server
have all been commented out.