Initial check-in of scripts.
db_cleanup.php removes all OER content from the OERbit DB. It is a
very good idea to have a dump of the database before unleashing this
script on it. You won't normally need to use this script. It's mostly
here for preparing seed DB's for Github.

reset_admin_info.php is a basic script to reset the uid=1 username and
password.