Add deleted version cleanup functionality.
Enabled the deletion of records that refer to deleted versions of
nodes instead of just records that refer to deleted nodes. This way,
records for which there are valid nids but invalid vids are also
removed.

The default methods called are such that the sample content isn't
removed from the DB.

Various documentation cleanups were also done while we were at it.